refactor MixPanel hook
